+++ libstdc++-v3/include/bits/regex.tcc
@@ -1,6 +1,6 @@
 // class template regex -*- C++ -*-
 
-// Copyright (C) 2013-2024 Free Software Foundation, Inc.
+// Copyright (C) 2013-2025 Free Software Foundation, Inc.
 //
 // This file is part of the GNU ISO C++ Library.  This library is free
 // software; you can redistribute it and/or modify it under the
@@ -444,7 +444,7 @@ namespace __detail
 		      __num *= 10;
 		      __num += __traits.value(*__next++, 10);
 		    }
-		  if (0 <= __num && __num < this->size())
+		  if (0 <= __num && size_t(__num) < this->size())
 		    __output(__num);
 		}
 	      else
