+++ libstdc++-v3/testsuite/23_containers/unordered_set/hash_policy/71181.cc
@@ -1,4 +1,4 @@
-// Copyright (C) 2016-2024 Free Software Foundation, Inc.
+// Copyright (C) 2016-2025 Free Software Foundation, Inc.
 //
 // This file is part of the GNU ISO C++ Library.  This library is free
 // software; you can redistribute it and/or modify it under the
@@ -30,7 +30,7 @@ template<typename _USet>
     auto bkts = us.bucket_count();
     for (int nb_insert = 1; nb_insert <= threshold; ++nb_insert)
       {
-	if (nb_insert > nb_reserved)
+	if (size_t(nb_insert) > nb_reserved)
 	  {
 	    nb_reserved = bkts;
 	    us.reserve(nb_reserved);
